Add a Redis strategy for long polling
This strategy uses Redis PUB/SUB to communicate messages to subscribers
in different servers. The idea is to connect multiple `Basic` strategies
with Redis pub/sub.

When a new message is published, the server that receives the request
will forward it to all other servers with Redis. When each server gets
the message, they will `Publish` it to the clients they have connected
(the ones in their own basic strategy).