Adding tests for multiple Redis strategies
This test case simulates multipler servers coordinated by Redis Pub/Sub.
When a server gets a subscribe request, it'll keep the connection open
and wait for a redis `Receive` event. Once the event happens, it'll
forward the message to the subscribers using a Basic strategy.

The `publisher` will send the redis `PUBLISH` command, that will alert
all the servers about the event.

In this test case, we have 1 server with two subscribers that gets
alerts from one publisher with no subscribers. We test that, even if the
publisher has no idea that the two connections are subscribed, when it
receives a publish request both subscribers will be alerted.